Extract isNegZero to own file
