+++ lib/samsam.js
@@ -6,17 +6,7 @@ var isArguments = require("./is-arguments");
 var isDate = require("./is-date");
 var isElement = require("./is-element");
 var isNaN = require("./is-nan");
-
-
-/**
- * @name samsam.isNegZero
- * @param Object value
- *
- * Returns ``true`` if ``value`` is ``-0``.
- */
-function isNegZero(value) {
-    return value === 0 && 1 / value === -Infinity;
-}
+var isNegZero = require("./is-neg-zero");
 
 /**
  * @name samsam.equal
